// src/app/core/models/customer.ts
export interface Customer {
  // customerId: number;
  customerName: string | null;
  mobileNo: string | null;
  email: string | null;
  address: string | null;
  city: string | null;
  state: string | null;
  gstIn: string | null;
  createdDate: string | null;
  updatedDate: string | null;
  isActive: boolean | null;
}